Changes for stage 1.
Update the replaceCard method in the Hand class to maintain order of
cards in hand after a card is replaced, removed several magic constants,
added Javadoc to methods which were missing it. Renamed the "cards"
variable in the Hand class to "hand" to indicate that it represented a
hand of cards, not just a random grouping of cards.